pyproject.toml
[project]
name = "floravoice-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.3"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.135.1",
    "pyjwt>=2.12.1",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]

[tool.fastapi]
entrypoint = "app.main:app"
